Missing non-Z timeseries in new CRS
