+++ src/components/textures/ProjectionTexture.ts
@@ -125,6 +125,24 @@ export function SetReprojectionTexture(dimArrays: Array<number>[]){
     useGlobalStore.setState({remapTexture});
 }
 
+export function sampleCRS(tex: THREE.DataTexture, u:number, v:number): [THREE.Vector2, boolean] {
+  // Samples an array given UVs
+  const { data, width, height } = tex.image;
+  if (!data) return [new THREE.Vector2(u, v), true];
+
+  const x = Math.floor(u * (width - 1));
+  const y = Math.floor(v * (height - 1));
+
+  const idx = (y * width + x) * 4; // RGBA
+  const newU = THREE.DataUtils.fromHalfFloat(data[idx + 0])
+  const newV = THREE.DataUtils.fromHalfFloat(data[idx + 1])
+  const valid = THREE.DataUtils.fromHalfFloat(data[idx + 2])
+  return [
+    new THREE.Vector2(newU,newV),
+    valid > 0.5
+  ];
+}
+
 export function reproject(resolution: number = 256){
     const {nativeCRS, destCRS, plotType} = usePlotStore.getState()
     if (!nativeCRS || !destCRS) return; // This shouldn't trigger as the button will be disabled for this same condition
